windows/requirements.txt
google-genai
python-dotenv
Pillow
pyperclip
